+++ R/get_zoltar_truth.R
@@ -107,7 +107,7 @@ preprocess_truth_for_zoltar <- function(target, issue_date = NULL){
 #'
 #' @export
 #'
-save_truth_for_zoltar <- function(save_location="./data-truth"){
+save_truth_for_zoltar <- function(save_location = "./data-truth"){
   
   df_cum_death <- covidData::preprocess_truth_for_zoltar("Cumulative Deaths")
   df_inc_death <- covidData::preprocess_truth_for_zoltar("Incident Deaths")